DEV: Remove buffered rendering from watched words
This is another refactoring in the multi-step process to remove all uses
of our custom Render Buffer.

Previous commit: bb31e7f5b6150e7d30461cdeef9b537f43f3caee in the series.

This commit affects the display of the Webhooks Deliver Status column. It
is just a refactor and does not change any functionality.