DEV: Remove buffered rendering from backup upload progress
This is another refactoring in the multi-step process to remove all uses
of our custom Render Buffer.

Previous commit: 54e4559aea2ff534fcd7b4117251d3eef1e85da1 in this
series.

This commit affects the display of the backup upload progress bar. It is
just a refactor and does not change any functionality.