DEV: Remove buffered rendering from admin backup logs
This is another refactoring in the multi-step process to remove all uses
of our custom Render Buffer.

Previous commit: da66950cf523adddf707f5b6768e1a189ffb40bd in this
series.

This commit affects the display of the backup logs. It is just a
refactor and does not change any functionality.